+++ rmatch/src/test/java/no/rmz/rmatch/semantics/LineAnchorSemanticsTest.java
@@ -132,11 +132,6 @@ public class LineAnchorSemanticsTest {
     }
 
     @Override
-    public String getCurrentRestString() {
-      return text.substring(Math.min(currentPos + 1, text.length()));
-    }
-
-    @Override
     public String getString(final int start, final int stop) {
       return text.substring(start, stop);
     }